[toa] add support for 64 bit pointers.

+++ src/tbm_bufmgr_android.c
@@ -205,7 +205,7 @@ _android_bo_handle(tbm_bufmgr_android bufmgr_android, tbm_bo_android bo_android,
 	switch (device) {
 	case TBM_DEVICE_DEFAULT:
 	case TBM_DEVICE_2D:
-		bo_handle.u32 = (uint32_t)bo_android->handler;
+		bo_handle.u64 = (uintptr_t)bo_android->handler;
 		break;
 	case TBM_DEVICE_CPU:
 		if (!bo_android->pBase) {
